Cleanup StateOps add state_machine.py
